# ARROW-3893: [C++] Improve adaptive int builder performance