@pitrou 2018-11-27
Benchmark of int width detection (AMD Ryzen 7, gcc 7.3.0):
```
--------------------------------------------------------------------------------
Benchmark                                         Time           CPU Iterations
--------------------------------------------------------------------------------
BM_DetectUIntWidthNoNulls/min_time:1.000         15 us         15 us      94140    38.057GB/s
BM_DetectUIntWidthNulls/min_time:1.000           36 us         36 us      38588   15.3941GB/s
BM_DetectIntWidthNoNulls/min_time:1.000          18 us         18 us      79188    31.413GB/s
BM_DetectIntWidthNulls/min_time:1.000            41 us         41 us      34523   13.6286GB/s
```