@pitrou 2018-11-27
Adaptive int builder benchmarks:
* before:
```
------------------------------------------------------------------------------------------------
Benchmark                                                         Time           CPU Iterations
------------------------------------------------------------------------------------------------
BM_BuildAdaptiveIntNoNulls/repeats:2                          12244 us      12234 us         54   5.10847GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2                          12293 us      12284 us         54   5.08758GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_mean                     12269 us      12259 us         54   5.09803GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_median                   12269 us      12259 us         54   5.09803GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_stddev                      35 us         36 us         54   15.1272MB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3              45940 us      45914 us         15   1.36121GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3              44678 us      44653 us         15   1.39963GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3              45632 us      45603 us         15   1.37047GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_mean         45417 us      45390 us         15    1.3771GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_median       45632 us      45603 us         15   1.37047GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_stddev         658 us        657 us         15   20.5352MB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2                         23285 us      23269 us         30   5.37198GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2                         23287 us      23272 us         30   5.37118GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2_mean                    23286 us      23270 us         30   5.37158GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2_median                  23286 us      23270 us         30   5.37158GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2_stddev                      2 us          2 us         30   597.548kB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2             46421 us      46396 us         14   1.34707GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2             46596 us      46571 us         14     1.342GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2_mean        46508 us      46483 us         14   1.34453GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2_median      46508 us      46483 us         14   1.34453GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2_stddev        124 us        124 us         14   3.66811MB/s
```

* after:
```
------------------------------------------------------------------------------------------------
Benchmark                                                         Time           CPU Iterations
------------------------------------------------------------------------------------------------
BM_BuildAdaptiveIntNoNulls/repeats:2                           9506 us       9495 us         74   6.58249GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2                           9527 us       9520 us         74   6.56462GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_mean                      9517 us       9508 us         74   6.57355GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_median                    9517 us       9508 us         74   6.57355GB/s
BM_BuildAdaptiveIntNoNulls/repeats:2_stddev                      15 us         18 us         74   12.9434MB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3              29520 us      29499 us         24   2.11864GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3              30365 us      30345 us         24    2.0596GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3              29487 us      29470 us         24   2.12075GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_mean         29791 us      29771 us         24   2.09966GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_median       29520 us      29499 us         24   2.11864GB/s
BM_BuildAdaptiveIntNoNullsScalarAppend/repeats:3_stddev         498 us        497 us         24   35.5477MB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2                         18526 us      18513 us         37   6.75184GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2                         18647 us      18635 us         37   6.70782GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2_mean                    18586 us      18574 us         37   6.72983GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2_median                  18586 us      18574 us         37   6.72983GB/s
BM_BuildAdaptiveUIntNoNulls/repeats:2_stddev                     86 us         86 us         37   31.8727MB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2             29618 us      29600 us         24   2.11141GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2             29992 us      29975 us         24   2.08503GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2_mean        29805 us      29787 us         24   2.09822GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2_median      29805 us      29787 us         24   2.09822GB/s
BM_BuildAdaptiveUIntNoNullsScalarAppend/repeats:2_stddev        264 us        265 us         24   19.1034MB/s
```