@pitrou 2018-11-27
Dictionary builder benchmarks:
* before:
```
-----------------------------------------------------------------------------------------------
Benchmark                                                        Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_BuildInt64DictionaryArrayRandom/repeats:2                   165 us        165 us       4178   462.219MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2                   165 us        165 us       4178   463.788MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean              165 us        165 us       4178   463.003MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_median            165 us        165 us       4178   463.003MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_stddev              0 us          0 us       4178   1.10959MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2               154 us        154 us       4365   495.413MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2               158 us        158 us       4365   484.158MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean          156 us        156 us       4365   489.786MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_median        156 us        156 us       4365   489.786MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_stddev          3 us          3 us       4365   7.95881MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                  158 us        157 us       4451   484.478MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                  157 us        157 us       4451   486.471MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean             157 us        157 us       4451   485.475MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_median           157 us        157 us       4451   485.475MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_stddev             0 us          0 us       4451   1.40936MB/s
BM_BuildStringDictionaryArray/repeats:2                        388 us        388 us       1819   257.354MB/s
BM_BuildStringDictionaryArray/repeats:2                        388 us        388 us       1819   257.319MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                   388 us        388 us       1819   257.337MB/s
BM_BuildStringDictionaryArray/repeats:2_median                 388 us        388 us       1819   257.337MB/s
BM_BuildStringDictionaryArray/repeats:2_stddev                   0 us          0 us       1819   25.7869kB/s
```

* after:
```
-----------------------------------------------------------------------------------------------
Benchmark                                                        Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_BuildInt64DictionaryArrayRandom/repeats:2                   107 us        106 us       6431   716.536MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2                   105 us        105 us       6431   725.463MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean              106 us        106 us       6431   720.999MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_median            106 us        106 us       6431   720.999MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_stddev              1 us          1 us       6431   6.31234MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2                99 us         98 us       6953   775.034MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2                98 us         98 us       6953   781.114MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean           98 us         98 us       6953   778.074MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_median         98 us         98 us       6953   778.074MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_stddev          1 us          1 us       6953   4.29916MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                   98 us         98 us       6837   781.993MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                   98 us         98 us       6837   775.452MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean              98 us         98 us       6837   778.723MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_median            98 us         98 us       6837   778.723MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_stddev             1 us          1 us       6837   4.62541MB/s
BM_BuildStringDictionaryArray/repeats:2                        276 us        276 us       2531    362.76MB/s
BM_BuildStringDictionaryArray/repeats:2                        275 us        275 us       2531   363.518MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                   276 us        275 us       2531   363.139MB/s
BM_BuildStringDictionaryArray/repeats:2_median                 276 us        275 us       2531   363.139MB/s
BM_BuildStringDictionaryArray/repeats:2_stddev                   0 us          0 us       2531   549.271kB/s
```